Fix e2e tests to work.
The problem was that Jest uses the source code and not the production code, but in the configuration we had the entities defined to be looked up from the dist directory.
Tried some other ways to overwrite it but that seems to be a bit hard. This seems to work nicely though - now we basically have the typeorm configuration in its own file so that
we can easily run the migrations with typeorm, but in the app module we get the data in that configuration manually so that we can overwrite the entities easily.

This still isn't perfect. I think it'd be better to have a dedicated database for tests, but this works for now.